Apply moving offset for residuals when time series are short (#12)
Previously, the set of non-NA residuals dropped back down to 1 when
a time series reached `period_length`-observations which lead to
collapsed prediction intervals. Now, we keep using residuals from
the initial observations until the time series has length of at
least two periods. This is not ideal as the early residuals are not
representative of the seasonal effects, but better than collapsed
prediction intervals.